[기능 구현] 미작성 리뷰 조회 api 구현

See merge request s11-fintech-finance-sub1/S11P21A601!294

+++ backend/franchise/src/main/java/kr/co/olivepay/franchise/repository/LikeRepository.java
@@ -10,7 +10,7 @@ import kr.co.olivepay.franchise.entity.Like;
 @Repository
 public interface LikeRepository extends JpaRepository<Like, Long> {
 
-	Integer countByFranchiseId(Long FranchiseId);
+	Long countByFranchiseId(Long FranchiseId);
 	Boolean existsByMemberIdAndFranchiseId(Long memberId, Long FranchiseId);
 
 	List<Like> getAllByMemberId(Long memberId);
